# backend/openapi.yaml
openapi: 3.1.0
info:
  title: LFMSH Bank API
  description: API for the LFMSH Bank System
  version: 0.1.0
  contact:
    name: LFMSH Team
    email: idutvuk@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
...
  /api/v1/users/:
    get:
      tags:
      - users
      summary: Read Users
      description: Retrieve users. Regular users can only access their own info and
        their party.
      operationId: read_users_api_v1_users__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserListItem'
                title: Response Read Users Api V1 Users  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - users
      summary: Create User
      description: Create new user. Only accessible to superusers.
      operationId: create_user_api_v1_users__post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/me:
    get:
      tags:
      - users
      summary: Read User Me
      description: Get current user.
      operationId: read_user_me_api_v1_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/users/{user_id}:
    get:
      tags:
      - users
      summary: Read User By Id
      description: Get a specific user by id.
      operationId: read_user_by_id_api_v1_users__user_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - users
      summary: Update User
      description: Update a user. Only accessible to superusers.
      operationId: update_user_api_v1_users__user_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CounterSchema:
      properties:
        counter_name:
          type: string
          title: Counter Name
        value:
          type: integer
          title: Value
        max_value:
          type: integer
          title: Max Value
      type: object
      required:
      - counter_name
      - value
      - max_value
      title: CounterSchema
      description: Schema for user counters
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    User:
      properties:
        username:
          type: string
          title: Username
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
          default: true
        id:
          type: integer
          title: Id
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        middle_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Middle Name
        party:
          type: integer
          title: Party
          default: 0
        grade:
          type: integer
          title: Grade
          default: 0
        is_staff:
          type: boolean
          title: Is Staff
          default: false
        is_superuser:
          type: boolean
          title: Is Superuser
          default: false
        balance:
          type: number
          title: Balance
          default: 0
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        staff:
          type: boolean
          title: Staff
          default: false
        expected_penalty:
          type: number
          title: Expected Penalty
          default: 0
        counters:
          items:
            $ref: '#/components/schemas/CounterSchema'
          type: array
          title: Counters
          default: []
        avatar:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar
      type: object
      required:
      - username
      - id
      - created_at
      title: User
    UserCreate:
      properties:
        username:
          type: string
          title: Username
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
          default: true
        password:
          type: string
          title: Password
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        middle_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Middle Name
        party:
          anyOf:
          - type: integer
          - type: 'null'
          title: Party
          default: 0
        grade:
          anyOf:
          - type: integer
          - type: 'null'
          title: Grade
          default: 0
        is_staff:
          type: boolean
          title: Is Staff
          default: false
        is_superuser:
          type: boolean
          title: Is Superuser
          default: false
      type: object
      required:
      - username
      - password
      title: UserCreate
    UserListItem:
      properties:
        id:
          type: integer
          title: Id
        username:
          type: string
          title: Username
        name:
          type: string
          title: Name
        party:
          type: integer
          title: Party
        staff:
          type: boolean
          title: Staff
        balance:
          type: number
          title: Balance
      type: object
      required:
      - id
      - username
      - name
      - party
      - staff
      - balance
      title: UserListItem
    UserUpdate:
      properties:
        username:
          type: string
          title: Username
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
          default: true
        password:
          anyOf:
          - type: string
          - type: 'null'
          title: Password
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        middle_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Middle Name
        party:
          anyOf:
          - type: integer
          - type: 'null'
          title: Party
        grade:
          anyOf:
          - type: integer
          - type: 'null'
          title: Grade
        is_staff:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Staff
        is_superuser:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Superuser
      type: object
      required:
      - username
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/jwt/create/
servers:
- url: http://localhost:8000
  description: Development server
- url: https://api.lfmsh-bank.com
  description: Production server
security:
- bearerAuth: []